Fix bugs, improve Redis logic, and add test coverage
- Fix queue name validation (was using && instead of ||, missing char check)
- Fix unsafe slice indexing in pop/receive message with safe get() + unwrap_or
- Fix InvalidValue error format string (wrong positional args {0:?} x3 → {0:?}/{1:?}/{2:?})
- Fix typos: BugCreatingRandonValue → BugCreatingRandomValue, stray backtick in error msg
- Fix create_queue atomicity: include SADD in the MULTI/EXEC pipeline with HSETNX ops
- Fix set_queue_attributes: clean up awkward reference-to-temporary pipe pattern
- Fix receiveMessage.lua: replace deprecated ZRANGEBYSCORE with ZRANGE BYSCORE, add nil body guard before side effects
- Add getQueueAttributes.lua: atomic TIME + HMGET + ZCARD + ZCOUNT via single Lua script
- Change maxsize from i32 to i64 for consistency with Lua integer return values
- Add 16 new tests covering name validation, maxsize boundaries, message size limits,
  attribute counters, partial updates, phantom message guard, and error cases

+++ src/trait.rs
@@ -30,7 +30,7 @@ pub trait RsmqConnection {
         qname: &str,
         hidden: Option<Duration>,
         delay: Option<Duration>,
-        maxsize: Option<i32>,
+        maxsize: Option<i64>,
     ) -> impl Future<Output = RsmqResult<()>> + Send;
 
     /// Deletes a message from the queue.
@@ -125,7 +125,7 @@ pub trait RsmqConnectionSync {
         qname: &str,
         hidden: Option<Duration>,
         delay: Option<Duration>,
-        maxsize: Option<i32>,
+        maxsize: Option<i64>,
     ) -> RsmqResult<()>;
 
     /// Deletes a message from the queue.
